IQueueObserver.notify(int,int)

+++ api/src/main/java/ru/kwanza/jeda/api/internal/IQueueObserver.java
@@ -4,5 +4,5 @@ package ru.kwanza.jeda.api.internal;
  * @author Guzanov Alexander
  */
 public interface IQueueObserver {
-    public void notifyChange(long queueSize, long delta);
+    public void notifyChange(int queueSize, int delta);
 }
