+++ core/src/test/java/ru/kwanza/jeda/core/queue/TestPriorityEventQueue.java
@@ -110,10 +110,10 @@ public abstract class TestPriorityEventQueue extends TestCase {
 
     public void testObserver() throws SinkException, SourceException {
         IQueue queue = createQueue();
-        final ArrayList<Long> queueSize = new ArrayList<Long>();
-        final ArrayList<Long> delta = new ArrayList<Long>();
+        final ArrayList<Integer> queueSize = new ArrayList<Integer>();
+        final ArrayList<Integer> delta = new ArrayList<Integer>();
         queue.setObserver(new IQueueObserver() {
-            public void notifyChange(long s, long d) {
+            public void notifyChange(int s, int d) {
                 queueSize.add(s);
                 delta.add(d);
             }
